Create NavigationLayer+View
