+++ Sources/ArcGISToolkit/Common/NavigationLayer/NavigationLayer.swift
@@ -214,51 +214,3 @@ extension NavigationLayer {
         return copy
     }
 }
-
-struct NavigationLayerHeaderBackground: PreferenceKey {
-    static let defaultValue: Color? = nil
-    
-    static func reduce(value: inout Color?, nextValue: () -> Color?) {
-        value = nextValue()
-    }
-}
-
-struct NavigationLayerTitle: PreferenceKey {
-    static let defaultValue: String? = nil
-    
-    static func reduce(value: inout String?, nextValue: () -> String?) {
-        value = nextValue()
-    }
-}
-
-struct NavigationLayerSubtitle: PreferenceKey {
-    static let defaultValue: String? = nil
-    
-    static func reduce(value: inout String?, nextValue: () -> String?) {
-        value = nextValue()
-    }
-}
-
-extension View {
-    /// Sets a header background color for the navigation layer destination.
-    /// - Parameter color: The color for the navigation layer destination.
-    func navigationLayerHeaderBackground(_ color: Color) -> some View {
-        preference(key: NavigationLayerHeaderBackground.self, value: color)
-    }
-    
-    /// Sets a title for the navigation layer destination.
-    /// - Parameters:
-    ///   - title: The title for the navigation layer destination.
-    func navigationLayerTitle(_ title: String) -> some View {
-        preference(key: NavigationLayerTitle.self, value: title)
-    }
-    
-    /// Sets a title and subtitle for the navigation layer destination.
-    /// - Parameters:
-    ///   - title: The title for the navigation layer destination.
-    ///   - subtitle: The subtitle for the navigation layer destination.
-    func navigationLayerTitle(_ title: String, subtitle: String) -> some View {
-        preference(key: NavigationLayerTitle.self, value: title)
-            .preference(key: NavigationLayerSubtitle.self, value: subtitle)
-    }
-}
